Revert "injecting an warning/error to test ci configuration"
This reverts commit 0edbf86a44eb6657517296559a14e578bec30e09.

+++ src/cmdargs.cpp
@@ -170,7 +170,7 @@ bool Config::ParseCmdLine(int argc, const _TCHAR* const* argv)
 
     // Switch
     // which can be specified multiple times:
-    m_verbose = res.HasSwitch(verboseSwitch);
+    m_verbose = static_cast<int>(res.HasSwitch(verboseSwitch));
 
     // Arguments
     // the typical case:
